chore: move core:parcelize to shared:screens
