+++ shared/screens/src/androidMain/kotlin/ly/david/musicsearch/shared/screens/CommonParcelize.kt
@@ -1,4 +1,4 @@
-package ly.david.musicsearch.core.parcelize
+package ly.david.musicsearch.shared.screens
 
 /**
  * Copied from https://github.com/realityexpander/NoteAppKMM
