+++ shared/screens/src/commonMain/kotlin/ly/david/musicsearch/shared/screens/Screens.kt
@@ -2,7 +2,6 @@ package ly.david.musicsearch.shared.screens
 
 import com.slack.circuit.runtime.screen.Screen
 import ly.david.musicsearch.core.models.network.MusicBrainzEntity
-import ly.david.musicsearch.core.parcelize.CommonParcelize
 
 @CommonParcelize
 data class SearchScreen(
