+++ ui/common-legacy/src/main/java/ly/david/ui/commonlegacy/EntitiesByEntityViewModel.kt
@@ -16,6 +16,9 @@ import ly.david.musicsearch.core.models.listitem.ListItemModel
 import ly.david.musicsearch.core.models.network.MusicBrainzEntity
 import ly.david.musicsearch.domain.base.usecase.GetEntitiesByEntity
 
+// TODO: it's the worst thing to not have this abstraction
+//  we have to overfill it with properties such as isRemote (only used by *ByCollectionViewModel)
+//  because we need to account for all usecases
 abstract class EntitiesByEntityViewModel<LI : ListItemModel>(
     private val entity: MusicBrainzEntity,
     private val getEntitiesByEntity: GetEntitiesByEntity<LI>,
